Export Dialog: Add ui enable/disable for options in export dialog, restructure options framework

+++ src/tilemap_read.h
@@ -5,4 +5,4 @@
 
 #include <glib.h>
 
-int tilemap_read(const gchar *, int);
+int tilemap_read(const gchar *, uint16_t);
